+++ packages/react-ui/src/localization/es.js
@@ -97,6 +97,10 @@ const locales = {
         clear: 'Limpiar',
         selectedItems: '{items, plural, one {# seleccionado} other {# seleccionados}}',
         allSelected: 'Todos seleccionados'
+      },
+      chartLegend: {
+        next: 'Página siguiente',
+        prev: 'Página anterior'
       }
     }
   }